Move getCVData to its own folder
